update[EX1] updated lasdtest folder with new tests from Mogavero

+++ exercise1/zlasdtest/container/container.hpp
@@ -8,9 +8,9 @@
 
 // Container member functions!
 
-void Empty(unsigned long&, unsigned long&, const lasd::Container&, bool);
+void Empty(uint&, uint&, const lasd::Container&, bool);
 
-void Size(unsigned long&, unsigned long&, const lasd::Container&, bool, unsigned long);
+void Size(uint&, uint&, const lasd::Container&, bool, ulong);
 
 /* ************************************************************************** */
