+++ exercise1/zlasdtest/exercise1a/simpletest.cpp
@@ -21,8 +21,8 @@ using namespace std;
 /* ************************************************************************** */
 
 // TODO: Uncomment after Vector implements
-// void stestVectorInt(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVectorInt(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of Vector<int> Test:" << endl;
 //   try {
 //     {
@@ -82,8 +82,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestVectorDouble(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVectorDouble(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of Vector<double> Test:" << endl;
 //   try {
 //     lasd::SortableVector<double> vec(3);
@@ -111,8 +111,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestVectorString(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVectorString(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of Vector<string> Test:" << endl;
 //   try {
 //     lasd::SortableVector<string> vec(2);
@@ -163,8 +163,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestVector(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVector(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   stestVectorInt(loctestnum, loctesterr);
 //   stestVectorDouble(loctestnum, loctesterr);
 //   stestVectorString(loctestnum, loctesterr);
@@ -176,8 +176,8 @@ using namespace std;
 /* ************************************************************************** */
 
 // TODO: Uncomment after List implementation
-// void stestListInt(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestListInt(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of List<int> Test:" << endl;
 //   try {
 //     lasd::List<int> lst;
@@ -262,8 +262,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestListDouble(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestListDouble(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of List<double> Test:" << endl;
 //   try {
 //     lasd::List<double> lst;
@@ -299,8 +299,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestListString(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestListString(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of List<string> Test:" << endl;
 //   try {
 //     lasd::List<string> lst;
@@ -345,8 +345,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestList(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestList(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   stestListInt(loctestnum, loctesterr);
 //   stestListDouble(loctestnum, loctesterr);
 //   stestListString(loctestnum, loctesterr);
@@ -358,8 +358,8 @@ using namespace std;
 /* ************************************************************************** */
 
 // TODO: Uncomment code after crossed Vector-List implementation
-// void stestVectorListInt(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVectorListInt(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of Vector/List<int> Test:" << endl;
 //   try {
 //     lasd::SortableVector<int> vec(3);
@@ -371,6 +371,8 @@ using namespace std;
 //     InsertAtFront(loctestnum, loctesterr, lst, true, 1);
 //     InsertAtFront(loctestnum, loctesterr, lst, true, 0);
 //     InsertAtFront(loctestnum, loctesterr, lst, true, -1);
+
+//     EqualLinear(loctestnum, loctesterr, vec, lst, true);
 
 //     lasd::SortableVector<int> copvec(lst);
 //     EqualVector(loctestnum, loctesterr, vec, copvec, true);
@@ -391,8 +393,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestVectorListDouble(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVectorListDouble(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of Vector/List<double> Test:" << endl;
 //   try {
 //     lasd::SortableVector<double> vec(3);
@@ -404,6 +406,8 @@ using namespace std;
 //     InsertAtBack(loctestnum, loctesterr, lst, true, -0.5);
 //     InsertAtBack(loctestnum, loctesterr, lst, true, 0.0);
 //     InsertAtBack(loctestnum, loctesterr, lst, true, 0.5);
+
+//     EqualLinear(loctestnum, loctesterr, vec, lst, true);
 
 //     lasd::SortableVector<double> copvec(lst);
 //     EqualVector(loctestnum, loctesterr, vec, copvec, true);
@@ -424,8 +428,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestVectorListString(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVectorListString(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   cout << endl << "Begin of Vector/List<string> Test:" << endl;
 //   try {
 //     lasd::SortableVector<string> vec(3);
@@ -437,6 +441,8 @@ using namespace std;
 //     InsertAtFront(loctestnum, loctesterr, lst, true, string("B"));
 //     InsertAtBack(loctestnum, loctesterr, lst, true, string("C"));
 //     InsertAtFront(loctestnum, loctesterr, lst, true, string("A"));
+
+//     EqualLinear(loctestnum, loctesterr, vec, lst, true);
 
 //     lasd::SortableVector<string> copvec(lst);
 //     EqualVector(loctestnum, loctesterr, vec, copvec, true);
@@ -465,8 +471,8 @@ using namespace std;
 //   testerr += loctesterr;
 // }
 
-// void stestVectorList(unsigned long & testnum, unsigned long & testerr) {
-//   unsigned long loctestnum = 0, loctesterr = 0;
+// void stestVectorList(uint & testnum, uint & testerr) {
+//   uint loctestnum = 0, loctesterr = 0;
 //   stestVectorListInt(loctestnum, loctesterr);
 //   stestVectorListDouble(loctestnum, loctesterr);
 //   stestVectorListString(loctestnum, loctesterr);
@@ -477,10 +483,10 @@ using namespace std;
 
 /* ************************************************************************** */
 
-void testSimpleExercise1A(unsigned long & testnum, unsigned long & testerr) {
+void testSimpleExercise1A(uint & testnum, uint & testerr) {
   // TODO: Uncomment to test Vector
   // stestVector(testnum, testerr);
-  // TODO: Uncomment to test List
+  // TODO: Uncomment to test Vector
   // stestList(testnum, testerr);
   // TODO: Uncomment to test Vector-List
   // stestVectorList(testnum, testerr);
