+++ exercise1/zlasdtest/exercise1b/fulltest.cpp
@@ -3,5 +3,5 @@
 
 /* ************************************************************************** */
 
-void testFullExercise1B(unsigned long &, unsigned long &) {
+void testFullExercise1B(uint &, uint &) {
 }
